chore: extract positions/portfolio from dashboard index (move-only)
