+++ src/routes/dashboard/index.ts
@@ -102,8 +102,10 @@ import { WebullAuth } from '../../infrastructure/webull/WebullAuth'
 import { WebullTokenClient } from '../../infrastructure/webull/WebullTokenClient'
 import { WebullTokenStateClient } from '../../trading/state/WebullTokenStateClient'
 import type { WebullTokenState } from '../../trading/state/WebullTokenStateDO'
-import { JST_FORMATTER, clampLimit, currencyOfSymbol, displaySymbol, esc, fmtJst, fmtNumber, fmtPct, fmtPctSigned, fmtPriceCcy, formatCooldown, inactiveTooltip, isSymbolInactive, jsonPretty, messageOf, parseCursor, safeJsonScript, unavailable } from './shared'
+import { clampLimit, currencyOfSymbol, displaySymbol, esc, fmtJst, fmtNumber, fmtPct, fmtPctSigned, fmtPriceCcy, inactiveTooltip, isSymbolInactive, jsonPretty, messageOf, parseCursor, safeJsonScript, unavailable } from './shared'
 import { type DashboardBindings, layout, loadKillSwitchState, renderLayout } from './layout'
+import { loadLatestStrategyPrices, pickFreshQuote, positionsBody } from './positions'
+import { type EquityRange, parseEquityRange, portfolioBody, renderPortfolioEquityChart, renderVixRegimeCell, safeLoadPortfolioSnapshots } from './portfolio'
 import { tradesBody } from './trades'
 import { cronBody, cronDecisionJson, formatRealizedPnl, loadDecisionRows, renderDecisionLadder, renderDecisionTable } from './cron'
 import { alertsBody, clampAlertLimit, parseAlertsQuery, parseEventTypeFilter, parseSeverityFilter } from './alerts'
@@ -113,6 +115,9 @@ import { MAX_CHART_DECISIONS, type SymbolChartData, type SymbolChartPoint, type 
 import { loadEquityCurve, renderOverviewTab } from './charts/equity'
 import { computePnlHistogram, computeTradeStats, loadDecisionBreakdown, loadTradePnls, renderQualityTab } from './charts/quality'
 export { safeJsonScript } from './shared'
+export { formatQuoteAsOf, pickFreshQuote } from './positions'
+export { parseEquityRange, renderLastRolledCell, renderVixRegimeCell } from './portfolio'
+export type { EquityRange } from './portfolio'
 export { localizeReason, renderChartDecisionTrace } from './cron'
 export type { DecisionRow } from './cron'
 export { renderAlertFilterPills } from './alerts'
@@ -2593,389 +2598,6 @@ function overviewBody(data: OverviewData): string {
   return buyingPowerBadge() + sections.join('')
 }
 
-/**
- * 各銘柄の「strategy が直近に判定で使った価格」を取得。
- * Yahoo daily bars から計算された `indicators.price` が
- * strategy_decision_log.price に書き出されているので、最新行を引く。
- *
- * Webull bridge が落ちて lastQuote が古い場合、こちらが新しければ
- * dashboard の現在値表示に採用される (pickFreshQuote で比較)。
- *
- * 実装: D1 の `(symbol, id)` 複合 index を活かして symbol 並列で
- * `ORDER BY id DESC LIMIT 1` を打つ。1 銘柄あたり 1 row のみ転送。
- */
-async function loadLatestStrategyPrices(
-  db: D1Database,
-  symbols: string[],
-): Promise<Map<string, { price: number; asOf: string }>> {
-  if (symbols.length === 0) return new Map()
-  const drizzle = createDb(db)
-  // 個別 symbol の失敗で全体を 500 にしないよう per-symbol で catch。
-  // strategy_decision_log がまだ空の銘柄や DB 一時的エラーは「Yahoo 価格なし」
-  // として扱い、Webull lastQuote にフォールバックさせる。
-  const entries = await Promise.all(
-    symbols.map(async (sym) => {
-      try {
-        const row = await drizzle
-          .select({
-            symbol: strategyDecisionLog.symbol,
-            price: strategyDecisionLog.price,
-            timestamp: strategyDecisionLog.timestamp,
-          })
-          .from(strategyDecisionLog)
-          .where(eq(strategyDecisionLog.symbol, sym))
-          .orderBy(desc(strategyDecisionLog.id))
-          .limit(1)
-        const r = row[0]
-        if (!r || r.price === null || r.price === undefined) return null
-        return [r.symbol, { price: r.price, asOf: r.timestamp }] as const
-      } catch {
-        return null
-      }
-    }),
-  )
-  return new Map(entries.filter((e): e is readonly [string, { price: number; asOf: string }] => e !== null))
-}
-
-/**
- * 表示用「現在値」の決定。dashboard が見せる現在値の source は
- * 2 系統あり、bridge 障害などで Webull snapshot が古くなる場合がある:
- *
- * - webull-snapshot: SymbolStateDO.lastQuote (Webull bridge の 5 分 cron)
- * - yahoo-bars: strategy_decision_log.price (Yahoo daily bars 経由、15 分 cron)
- *
- * 両方あれば asOf が新しい方を採用。strategy が判定に使う価格と表示が
- * 一致するのが UX 上の正なので、片方だけしか無い場合もそちらを採る。
- */
-interface ResolvedQuote {
-  price: number
-  source: string
-  asOf: string
-}
-
-export function pickFreshQuote(
-  webull: { price: number; source: string; asOf: string } | null,
-  yahoo: { price: number; asOf: string } | null,
-): ResolvedQuote | null {
-  if (webull === null && yahoo === null) return null
-  if (webull === null) return { price: yahoo!.price, source: 'yahoo-bars', asOf: yahoo!.asOf }
-  if (yahoo === null) return { price: webull.price, source: webull.source, asOf: webull.asOf }
-  const w = new Date(webull.asOf).getTime()
-  const y = new Date(yahoo.asOf).getTime()
-  // 不正な ISO は "より古い" 扱い: 有効な側があればそちらを採用、両方
-  // 不正なら webull にタイブレーク (既存挙動維持)。`y > w` だけだと
-  // w=NaN の時に false 評価で不正な webull を選んでしまう回帰がある。
-  const wValid = Number.isFinite(w)
-  const yValid = Number.isFinite(y)
-  const pickYahoo = yValid && (!wValid || y > w)
-  return pickYahoo
-    ? { price: yahoo.price, source: 'yahoo-bars', asOf: yahoo.asOf }
-    : { price: webull.price, source: webull.source, asOf: webull.asOf }
-}
-
-function positionsBody(
-  rows: Array<{ sym: string; state: SymbolState | null; error: string | null }>,
-  strategyPriceMap: Map<string, { price: number; asOf: string }>,
-  universe?: SymbolUniverse | null,
-): string {
-  if (rows.length === 0) return `<p class="muted">有効な銘柄がありません。</p>`
-  const tbody = rows
-    .map((r) => {
-      const inactive = isSymbolInactive(r.sym, universe)
-      const rowClass = inactive ? ' class="symbol-disabled-row"' : ''
-      const symbolClass = inactive ? ' class="symbol-disabled"' : ''
-      const titleAttr = inactive ? ` title="${esc(inactiveTooltip(r.sym, universe))}"` : ''
-      if (r.error !== null || r.state === null) {
-        return `<tr${rowClass}><td><a href="/dashboard/charts?tab=symbol&symbol=${encodeURIComponent(r.sym)}"${titleAttr} style="text-decoration:none"><span${symbolClass}>${esc(displaySymbol(r.sym, universe))}</span></a></td><td colspan="7" class="err">${esc(r.error ?? '状態取得不可')}</td></tr>`
-      }
-      const s = r.state
-      const pos = s.position
-      const webull = s.lastQuote
-        ? { price: s.lastQuote.price, source: s.lastQuote.source, asOf: s.lastQuote.asOf ?? s.lastQuote.fetchedAt }
-        : null
-      const yahoo = strategyPriceMap.get(s.symbol) ?? null
-      const quote = pickFreshQuote(webull, yahoo)
-      const pendingSide = s.pendingOrder?.side
-      const pnlPct =
-        pos !== null && quote !== null && pos.avgPrice > 0
-          ? ((quote.price - pos.avgPrice) / pos.avgPrice) * 100
-          : null
-      const pnlClass = pnlPct === null ? 'muted' : pnlPct >= 0 ? 'ok' : 'err'
-      const quoteCell = quote
-        ? `${fmtNumber(quote.price, 2)} <span class="muted" style="font-size:11px">(${esc(quote.source)}, ${esc(formatQuoteAsOf(quote.asOf))})</span>`
-        : '<span class="muted">—</span>'
-      // 銘柄リンクはチャート銘柄タブへ (trades / cron と同じ導線)。▼ は判定
-      // 絞り込み、⚙ は設定編集への補助リンク (#nav-links)。
-      return `<tr${rowClass}>
-        <td><a href="/dashboard/charts?tab=symbol&symbol=${encodeURIComponent(s.symbol)}"${titleAttr} style="text-decoration:none"><strong><span${symbolClass}>${esc(displaySymbol(s.symbol, universe))}</span></strong></a> <a href="/dashboard/cron?symbol=${encodeURIComponent(s.symbol)}" class="muted" title="この銘柄の判定だけに絞り込み" style="font-size:11px;text-decoration:none">▼</a> <a href="/dashboard/symbols/${encodeURIComponent(s.symbol)}/edit" class="muted" title="銘柄設定を編集" style="font-size:11px;text-decoration:none">⚙</a></td>
-        <td>${pos ? esc(pos.qty) : '<span class="muted">—</span>'}</td>
-        <td>${pos ? fmtNumber(pos.avgPrice, 2) : '<span class="muted">—</span>'}</td>
-        <td>${quoteCell}</td>
-        <td class="${pnlClass}">${pnlPct === null ? '—' : fmtNumber(pnlPct, 2) + '%'}</td>
-        <td>${pendingSide ? esc(pendingSide) : '<span class="muted">—</span>'}</td>
-        <td>${formatCooldown(s.cooldownUntil)}</td>
-        <td class="muted">${esc(fmtJst(s.updatedAt))}</td>
-      </tr>`
-    })
-    .join('')
-  return `<p class="muted" style="font-size:12px">
-    評価損益は未実現 (現在値 vs 平均取得単価)。実約定損益は
-    <a href="/dashboard/cron">/dashboard/cron</a> 「実 損益」列を参照。
-  </p>
-  <table>
-    <thead><tr>
-      <th>銘柄</th><th>数量</th><th>平均取得単価</th><th>現在値 (source, asOf)</th><th>評価損益</th>
-      <th>未約定</th><th>クールダウン</th><th>更新時刻</th>
-    </tr></thead>
-    <tbody>${tbody}</tbody>
-  </table>`
-}
-
-function portfolioBody(p: {
-  dailyStartEquity: number
-  dailyRealizedPnl: number
-  tradingDisabledUntil: string | null
-  lastRolledAt?: string | null
-  updatedAt: string
-}, vixRegime: VixRegime | null, equity?: {
-  snapshots: PortfolioEquitySnapshotRow[]
-  range: EquityRange
-}): string {
-  const drawdownPct =
-    p.dailyStartEquity > 0 ? (p.dailyRealizedPnl / p.dailyStartEquity) * 100 : null
-  const ddClass = drawdownPct === null ? 'muted' : drawdownPct >= 0 ? 'ok' : 'err'
-  const kill = p.tradingDisabledUntil
-  const lastRolledCell = renderLastRolledCell(p.lastRolledAt ?? null)
-  const vixCell = renderVixRegimeCell(vixRegime)
-  const summaryTable = `<table>
-    <tbody>
-      <tr><th>当日始値資産 (dailyStartEquity)</th><td>${fmtNumber(p.dailyStartEquity, 2)}</td></tr>
-      <tr><th>当日実現損益 (dailyRealizedPnl)</th><td class="${ddClass}">${fmtNumber(p.dailyRealizedPnl, 2)}</td></tr>
-      <tr><th>ドローダウン (drawdown)</th><td class="${ddClass}">${drawdownPct === null ? '—' : fmtNumber(drawdownPct, 2) + '%'}</td></tr>
-      <tr><th>取引停止解除時刻 (tradingDisabledUntil)</th><td>${kill ? `<span class="warn">${esc(fmtJst(kill))}</span>` : '<span class="ok">稼働中</span>'}</td></tr>
-      <tr><th>VIX レジーム (vixRegime)</th><td>${vixCell}</td></tr>
-      <tr><th>EOD ロールオーバー実行時刻 (lastRolledAt)</th><td>${lastRolledCell}</td></tr>
-      <tr><th>更新時刻 (updatedAt)</th><td class="muted">${esc(fmtJst(p.updatedAt))}</td></tr>
-    </tbody>
-  </table>`
-  const chartSection = equity ? renderPortfolioEquityChart(equity.snapshots, equity.range) : ''
-  return summaryTable + chartSection
-}
-
-/**
- * `?range=30d|90d|365d|all` の解釈。default は 90d (3 ヶ月で trend が読める粒度)。
- * 不正値は default に倒す。`all` は cap 内 (3650 件) で全件返し。
- */
-export type EquityRange = '30d' | '90d' | '365d' | 'all'
-
-export function parseEquityRange(value: string | undefined): EquityRange {
-  if (value === '30d' || value === '90d' || value === '365d' || value === 'all') return value
-  return '90d'
-}
-
-function equityRangeLimit(range: EquityRange): number {
-  if (range === '30d') return 30
-  if (range === '90d') return 90
-  if (range === '365d') return 365
-  return 3650
-}
-
-/**
- * `loadPortfolioEquitySnapshots` を try/catch で wrap。table 未 migration や
- * D1 エラー時は空配列で fallback → ページ自体は描画。チャート枠は "データ無し"
- * メッセージに置き換わる。
- */
-async function safeLoadPortfolioSnapshots(
-  db: D1Database,
-  range: EquityRange,
-): Promise<PortfolioEquitySnapshotRow[]> {
-  const opts: LoadPortfolioEquitySnapshotOptions = { limit: equityRangeLimit(range) }
-  try {
-    return await loadPortfolioEquitySnapshots(db, opts)
-  } catch (err) {
-    console.error(
-      JSON.stringify({
-        event: 'portfolio_equity_snapshot_load_failed',
-        error: err instanceof Error ? err.message : String(err),
-      }),
-    )
-    return []
-  }
-}
-
-/**
- * `/dashboard/portfolio` の総資産チャート。echarts inline JS で USD / JPY の
- * 2 ライン (片方 NULL は skip)。
- *
- * - X: snapshotAt の日付部分 (UTC date)
- * - Y: dailyStartEquity (通貨別)
- * - range tab で 30d / 90d / 365d / all 切替 (URL `?range=` を維持)
- */
-function renderPortfolioEquityChart(
-  snapshots: PortfolioEquitySnapshotRow[],
-  range: EquityRange,
-  basePath = '/dashboard/portfolio',
-): string {
-  const rangeTabs = renderEquityRangeTabs(range, basePath)
-  if (snapshots.length === 0) {
-    return `<h3 style="margin-top:24px">総資産チャート</h3>
-    ${rangeTabs}
-    <p class="muted">まだ roll-daily 実行履歴がありません。<code>/admin/portfolio/roll-daily</code> を実行すると、ここに時系列が描画されます。</p>`
-  }
-  const usdPoints: Array<{ date: string; value: number | null }> = []
-  const jpyPoints: Array<{ date: string; value: number | null }> = []
-  let hasUsd = false
-  let hasJpy = false
-  for (const row of snapshots) {
-    const date = (row.snapshotAt ?? '').slice(0, 10)
-    const usd =
-      typeof row.dailyStartEquityUsd === 'number' && Number.isFinite(row.dailyStartEquityUsd)
-        ? row.dailyStartEquityUsd
-        : null
-    const jpy =
-      typeof row.dailyStartEquityJpy === 'number' && Number.isFinite(row.dailyStartEquityJpy)
-        ? row.dailyStartEquityJpy
-        : null
-    if (usd !== null) hasUsd = true
-    if (jpy !== null) hasJpy = true
-    usdPoints.push({ date, value: usd })
-    jpyPoints.push({ date, value: jpy })
-  }
-  const initScript = `
-    document.addEventListener('DOMContentLoaded', function () {
-      if (typeof echarts === 'undefined') return;
-      var data = window.__equityChartData;
-      var dates = data.usd.map(function (p) { return p.date; });
-      var series = [];
-      if (data.hasUsd) {
-        series.push({
-          name: 'USD',
-          type: 'line',
-          data: data.usd.map(function (p) { return p.value; }),
-          connectNulls: false,
-          smooth: false,
-          lineStyle: { width: 2, color: '#1471a8' },
-          itemStyle: { color: '#1471a8' },
-        });
-      }
-      if (data.hasJpy) {
-        series.push({
-          name: 'JPY',
-          type: 'line',
-          yAxisIndex: data.hasUsd ? 1 : 0,
-          data: data.jpy.map(function (p) { return p.value; }),
-          connectNulls: false,
-          smooth: false,
-          lineStyle: { width: 2, color: '#b25000' },
-          itemStyle: { color: '#b25000' },
-        });
-      }
-      var yAxis = [{ type: 'value', name: 'USD', axisLabel: { formatter: '{value}' } }];
-      if (data.hasUsd && data.hasJpy) {
-        yAxis.push({ type: 'value', name: 'JPY', axisLabel: { formatter: '{value}' } });
-      } else if (!data.hasUsd && data.hasJpy) {
-        yAxis = [{ type: 'value', name: 'JPY', axisLabel: { formatter: '{value}' } }];
-      }
-      var chart = echarts.init(document.getElementById('portfolio-equity-chart'));
-      chart.setOption({
-        title: { text: '総資産 (dailyStartEquity) 時系列', left: 'center', textStyle: { fontSize: 14 } },
-        tooltip: { trigger: 'axis', valueFormatter: function (v) { return v == null ? '—' : Number(v).toFixed(2); } },
-        legend: { top: 24 },
-        grid: { left: 50, right: 20, top: 60, bottom: 40, containLabel: true },
-        xAxis: { type: 'category', data: dates },
-        yAxis: yAxis,
-        series: series,
-      });
-      window.addEventListener('resize', function () { chart.resize(); });
-    });
-  `
-  return `<h3 style="margin-top:24px">総資産チャート</h3>
-  ${rangeTabs}
-  <p class="muted" style="font-size:12px">
-    <code>PortfolioStateDO.dailyStartEquity</code> の roll-daily 時点スナップショット。
-    <code>/dashboard/charts?tab=overview</code> は <code>trade_journal.realized_pnl</code> の
-    累積で、こちらは口座総資産そのもの (cash + 保有時価)。USD / JPY を別軸でプロット。
-  </p>
-  <div id="portfolio-equity-chart" style="width:100%;height:320px;background:#fff;border:1px solid #d0d0d5;border-radius:6px;margin-top:12px"></div>
-  ${safeJsonScript('__equityChartData', { usd: usdPoints, jpy: jpyPoints, hasUsd, hasJpy })}
-  <script src="${ECHARTS_CDN}" defer></script>
-  <script>${initScript}</script>`
-}
-
-function renderEquityRangeTabs(active: EquityRange, basePath = '/dashboard/portfolio'): string {
-  const options: Array<{ id: EquityRange; label: string }> = [
-    { id: '30d', label: '30 日' },
-    { id: '90d', label: '90 日' },
-    { id: '365d', label: '365 日' },
-    { id: 'all', label: '全期間' },
-  ]
-  const links = options
-    .map((opt) => {
-      const cls = opt.id === active ? 'tab tab-active' : 'tab'
-      return `<a class="${cls}" href="${basePath}?range=${opt.id}">${opt.label}</a>`
-    })
-    .join(' ')
-  return `<div class="tab-strip" style="margin-top:12px">${links}</div>`
-}
-
-/**
- * VIX regime snapshot を bage 風に表示 (issue #196 3/3)。
- *
- *   - normal:   緑 (size 1.0、通常運用)
- *   - warning:  黄 (size 0.5、新規 BUY 縮小)
- *   - critical: 赤 (新規 BUY 全停止 / SELL は通常)
- *   - null:     灰 (snapshot 未生成、初回 cron tick 前 or DB 未配線)
- *
- * VIX 値そのものは snapshot table に持たないので regime ラベルのみ表示。
- * 値が必要なら strategy_decision_log の VIX reject reason を見る運用 (POC)。
- */
-export function renderVixRegimeCell(regime: VixRegime | null): string {
-  if (regime === null) {
-    return `<span class="muted">— (cron 未到達 or DB 未配線、fail-open で通常運用)</span>`
-  }
-  if (regime === 'critical') {
-    return `<span class="err">critical — 新規買い停止 (売却は通常)</span>`
-  }
-  if (regime === 'warning') {
-    return `<span class="warn">warning — 新規買いを縮小 (size scale 適用)</span>`
-  }
-  return `<span class="ok">normal — 通常運用</span>`
-}
-
-/**
- * Issue #140: `lastRolledAt` の経過時間で badge 色を切替。
- *  - null:       未実行 (muted)
- *  - <24h:       OK (ok)
- *  - 24h–48h:    warning (warn)
- *  - >=48h:      error  (err)
- *
- * EOD cron は毎日 22:00 UTC に走るので 24h 以内なら正常、48h 超は **2 日連続
- * miss** で要調査。閾値は `runStrategyCron.emitStaleRollWarningIfNeeded` の
- * 24h と一貫させている。
- */
-export function renderLastRolledCell(
-  lastRolledAt: string | null,
-  now: () => number = Date.now,
-): string {
-  if (lastRolledAt === null) {
-    return `<span class="warn">未実行 (EOD cron 未到達 or PORTFOLIO_STATE 未配線)</span>`
-  }
-  const ms = new Date(lastRolledAt).getTime()
-  if (!Number.isFinite(ms)) {
-    return `<span class="err">${esc(lastRolledAt)} (parse 不能)</span>`
-  }
-  const elapsedHours = (now() - ms) / 3_600_000
-  const formatted = esc(fmtJst(lastRolledAt))
-  const elapsedLabel = `${elapsedHours.toFixed(1)}h 前`
-  if (elapsedHours >= 48) {
-    return `<span class="err">${formatted} <small>(${esc(elapsedLabel)}, 48h 超 — EOD cron 要確認)</small></span>`
-  }
-  if (elapsedHours >= 24) {
-    return `<span class="warn">${formatted} <small>(${esc(elapsedLabel)}, 24h 超)</small></span>`
-  }
-  return `<span class="ok">${formatted} <small class="muted">(${esc(elapsedLabel)})</small></span>`
-}
-
 function configBody(
   global: Awaited<ReturnType<typeof loadGlobalConfigFrom>>,
   universe: Awaited<ReturnType<typeof loadSymbolUniverse>>,
@@ -3200,19 +2822,6 @@ const CONFIG_KEY_META: Record<string, ConfigKeyMeta> = {
     label: 'VIX 警戒時の建玉縮小率 (比率)',
     detail: 'VIX 警戒時 (warning ≤ VIX < critical) の発注数量倍率。0.5 = 半分に縮小。1.0 で縮小なし、0 で停止と同義。',
   },
-}
-
-/**
- * QuoteSnapshot.asOf (ISO) を JST の絶対表記 `MM/DD HH:MM JST` に。
- * 相対表記 (NN日前) は週末・場外で必ず古く見えてしまい「壊れている風」に
- * 誤読されやすいため、絶対時刻を出して「金曜引け」と一目で分かるようにする。
- */
-export function formatQuoteAsOf(asOf: string): string {
-  const d = new Date(asOf)
-  if (!Number.isFinite(d.getTime())) return '?'
-  const parts = JST_FORMATTER.formatToParts(d)
-  const pick = (type: string) => parts.find((p) => p.type === type)?.value ?? ''
-  return `${pick('month')}/${pick('day')} ${pick('hour')}:${pick('minute')} JST`
 }
 
 function formatConfigValue(v: unknown): string {
